refactor: the user should not have access to tasks inside the TaskManager
if the TaskManager method returns a reference to the task object to the user, it should be a reference to a new object with the data of the required object inside the manager. This way the user does not have access to the data inside the InMemoryTaskManager directly. Implement this by creating a copy constructor in each Task class so that each time the manager method is called, a separate object is created to return with the same data stored in the object inside TaskManager