refactor: the history has no limit on the number of tasks stored. if a task is deleted in the manager, it is also deleted in the viewing history
The history should be stored in a separately implemented data structure similar to LinkedHashMap. When calling any manager method that deletes a task, whether it is deleting a single task or all tasks of a certain type - they should also be deleted from the history. There are no repeated elements in the history, if you need to add a task to the history that is already stored in it, you need to delete the stored version and add a new one to the end of the list