fix: fix some bugs related to FileBackedManager methods.
To create an In Memory History Manager, you can use the Managers factory. To increase the readability of the code, it is better to create variables with clear names, and then pass them to the class constructor. It is not necessary to make the storages static, they must belong to a specific object. Remove useless BufferedReader.