feat: the task tracker will be able to prioritize tasks and check if they overlap in time of completion. Additionally several more tests must be written  for new functionality.
Add new fields for the Task class and its descendants: start Time and duration. Create the getPrioritizedTasks method, which will return tasks sorted by chronology, with the start time and duration specified. Add a method to check the intersection of tasks by execution time. If the tasks overlap, an error will be displayed. In some places of the code, replace the forEach loops with the API Stream. Cover all new functionality with tests. Divide the tests into classes that will inherit from the abstract Task Manager Test, which contains common tests for InMemoryTaskManagerTest and FileBackedTaskManager.