feat: file backed task manager should restore a prioritized list of tasks from a file, the principle of "program to an interface, not an implementation" should be followed, add the use of record instead of a local class
treeSetByTime should be restored from the file. In some places, when it is necessary to declare variables with the name of the interface rather than its specific implementation, it is the same with the types of variables returned by methods. In the updateEpicTime method, instead of the local Result class, you can use the record class.